use full 64-bits for displacements

+++ phf/src/map.rs
@@ -18,7 +18,7 @@ pub struct Map<K: 'static, V: 'static> {
     #[doc(hidden)]
     pub keys: [u64; 3],
     #[doc(hidden)]
-    pub disps: Slice<(u32, u32)>,
+    pub disps: Slice<(u64, u64)>,
     #[doc(hidden)]
     pub entries: Slice<(K, V)>,
 }
